Resolved CID 351642 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument this->m_playersinfo.size() to format specifier %u was expected to have type unsigned int but has type unsigned long

+++ src/world/Objects/ObjectMgr.cpp
@@ -484,7 +484,7 @@ void ObjectMgr::LoadPlayersInfo()
         while (result->NextRow());
         delete result;
     }
-    LogDetail("ObjectMgr : %u players loaded.", m_playersinfo.size());
+    LogDetail("ObjectMgr : %u players loaded.", static_cast<uint32_t>(m_playersinfo.size()));
 }
 
 PlayerInfo* ObjectMgr::GetPlayerInfoByName(const char* name)
